아침에 받은내용들 / 디버그하는 버릇

+++ Assets/FakeGame/Scripts/GamePlay/Enemy/EnemyController.cs
@@ -156,7 +156,7 @@ namespace Fake.Enemy
                     currentHP--;
                 }
 
-                GreenHpBar.fillAmount = currentHP / maxHP;
+                GreenHpBar.fillAmount = (float)currentHP / maxHP;
 
                 if (currentHP < 0)
                 {
